viscap/requirements.txt
cython==0.29.1
h5py==2.8.0
nltk==3.4
numpy==1.15.4
Pillow==5.3.0
pyyaml>=4.2b1
six==1.11.0
tensorboardX==1.2
tensorflow==1.12.0
torch==1.0.0
tqdm==4.28.1
